+++ lib/active_record_host_pool.rb
@@ -4,6 +4,7 @@ require 'active_record'
 require 'active_record/base'
 require 'active_record/connection_adapters/abstract_adapter'
 
+require 'active_record_host_pool/clear_query_cache_patch'
 require 'active_record_host_pool/connection_proxy'
 require 'active_record_host_pool/pool_proxy'
 require 'active_record_host_pool/connection_adapter_mixin'